Widen length-derived int arithmetic that overflows below the maxseqlength cap
The --maxseqlength cap (INT_MAX - buffer_headroom) makes a single length fit
int, but several sites multiply or double a length in int, overflowing well
below the cap and corrupting the heap or producing wrong values:

- chimera.cc: max_2x2_size = maxcandidates * maxqlen and maxalnlen = maxqlen +
  2*longest were computed in int before sizing the per-alignment vectors
  (overflow -> undersized resize -> OOB); the per-candidate index
  i * query_len + qpos was also int. Compute these in size_t / int64_t.
- msa.cc: the profile index profsize * position was computed in int before the
  size_t cast (overflow -> OOB into the profile vector). Widen the operands.
- unique.cc / kmerhash.cc: the kmer hash grows to 2 * seqlen; both the target
  and the doubling counter were int, so 2*seqlen wrapped negative (undersized
  hash -> OOB) and alloc *= 2 overflowed before reaching the target for
  sequences above ~1.07 Gnt. Carry size/alloc as int64_t and compute the
  target in 64-bit.
- search_exact.cc: nwscore = qseqlen * opt_match was an int*int product (signed
  overflow, UB) for long exact matches; compute in int64_t before the store.

All reachable only with very large single sequences (>~1 Gnt, mostly needing
--maxseqlength raised); behaviour is unchanged for normal inputs. Default and
--enable-debug builds are clean.

+++ src/search_exact.cc
@@ -121,7 +121,7 @@ auto add_hit(struct searchinfo_s * si, uint64_t seqno) -> void
 
       hp->count = 0;
 
-      hp->nwscore = static_cast<int>(si->qseqlen * opt_match);
+      hp->nwscore = static_cast<int>(static_cast<int64_t>(si->qseqlen) * opt_match);
       hp->nwdiff = 0;
       hp->nwgaps = 0;
       hp->nwindels = 0;
